refactor: split migration tunnel types into separate files
Move MigrationTunnelManager, migrationTunnel, and tunnelChannel (and
their tests) into dedicated files for clearer ownership boundaries.
